Rego policy engine
Adds an additional policy engine for using rego, and tests.

Supports filtering on agents, classes, facts, and if the provisioning mode is enabled

Support tracing on policy decisions, and a separate configuration entry to enable this on demand.  Defaults to false, enables on either a specific key or the debug log level being enabled.